# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/questions": {
      "get": {
        "operationId": "QuestionController_findAll",
        "summary": "List questions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      },
      "post": {
        "operationId": "QuestionController_create",
        "summary": "Create question",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      }
    },
    "/questions/{id}": {
      "get": {
        "operationId": "QuestionController_findOne",
        "summary": "Get question by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      },
      "patch": {
        "operationId": "QuestionController_update",
        "summary": "Update question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      },
      "delete": {
        "operationId": "QuestionController_remove",
        "summary": "Soft delete question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteQuestionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      }
    },
    "/questions/similar": {
      "post": {
        "operationId": "QuestionController_findSimilar",
        "summary": "Find similar questions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindSimilarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindSimilarResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      }
    },
    "/questions/bulk-delete": {
      "post": {
        "operationId": "QuestionController_bulkRemove",
        "summary": "Soft delete questions in bulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteQuestionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteQuestionsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      }
    },
    "/questions/{id}/restore": {
      "patch": {
        "operationId": "QuestionController_restore",
        "summary": "Restore soft deleted question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "questions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Interview App API",
    "description": "HTTP API contract for Interview App backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "QuestionExpectedConceptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "weight",
          "description"
        ]
      },
      "QuestionRedFlagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ]
          }
        },
        "required": [
          "id",
          "label",
          "severity"
        ]
      },
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "followUpQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expectedConcepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionExpectedConceptDto"
            }
          },
          "redFlags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionRedFlagDto"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "weight": {
            "type": "number"
          },
          "sampleGoodAnswer": {
            "type": "string"
          },
          "minimumPassScore": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "outputLanguage",
          "questionText",
          "followUpQuestions",
          "expectedConcepts",
          "redFlags",
          "difficulty",
          "weight",
          "minimumPassScore",
          "tags",
          "metadata",
          "createdAt",
          "updatedAt",
          "deleted"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Validation failed"
          },
          "path": {
            "type": "string",
            "example": "/questions/invalid-id"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "CreateQuestionDto": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "followUpQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expectedConcepts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "redFlags": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "weight": {
            "type": "number"
          },
          "sampleGoodAnswer": {
            "type": "string"
          },
          "minimumPassScore": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "questionText"
        ]
      },
      "FindSimilarDraftDto": {
        "type": "object",
        "properties": {
          "questionText": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          }
        }
      },
      "FindSimilarDto": {
        "type": "object",
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/FindSimilarDraftDto"
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 20
          },
          "excludeQuestionId": {
            "type": "string"
          }
        }
      },
      "SimilarQuestionMatchDto": {
        "type": "object",
        "properties": {
          "question": {
            "$ref": "#/components/schemas/QuestionResponseDto"
          },
          "score": {
            "type": "number"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "question",
          "score",
          "reasons"
        ]
      },
      "FindSimilarResponseDto": {
        "type": "object",
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimilarQuestionMatchDto"
            }
          }
        },
        "required": [
          "matches"
        ]
      },
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "followUpQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expectedConcepts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "redFlags": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "weight": {
            "type": "number"
          },
          "sampleGoodAnswer": {
            "type": "string"
          },
          "minimumPassScore": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "DeleteQuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "BulkDeleteQuestionsDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "BulkDeleteBlockedItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "questionText",
          "reason"
        ]
      },
      "BulkDeleteQuestionsResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "blocked": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkDeleteBlockedItemDto"
            }
          }
        },
        "required": [
          "deleted",
          "blocked"
        ]
      },
...
    }
  }
}